% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHotspots.R
\name{PlotHotspots}
\alias{PlotHotspots}
\title{Plots an object obtained with \code{DiffHotspots_n_k}}
\usage{
PlotHotspots(X, hotspots, order_extension = NULL, which.plot = NULL,
  rotation_angle = 0, eps_image = F)
}
\arguments{
\item{X}{- A \code{lpp} object representing a marked point pattern lying on a road network (\code{linnet} object)}

\item{hotspots}{- A set of differential risk hotspots obtained with the function \code{DiffHotspots_n_k}}

\item{order_extension}{- A natural number indicating a neighbourhood order to be used for constructing an extension of the differential risk hotspots. The summary is also given for the segments forming this extension}

\item{which.plot}{- A numeric vector indicating which differential risk hotspots to plot (according to the way they are ordered in \code{hotspots})}

\item{rotation_angle}{- A rotation angle (in degrees, from 0 to 180) to apply to the network (to improve visualization, if required). By default it is set to 0}

\item{eps_image}{- If set to \code{TRUE}, an .eps image is generated. By default it is set to \code{FALSE}}
}
\description{
This function plots a set of differential risk hotspots located along a linear network. An extension of the hotspots (including the kth order neighbourhs of the segments of the hotspots) is also plotted
}
\examples{
library(DRHotNet)
library(spatstat.core)
library(spdep)
library(raster)
library(maptools)
\donttest{
rel_probs_rear_end <- RelativeProbabilityNetwork(X = SampleMarkedPattern, 
lixel_length = 50, sigma = 100, mark = "Collision", category_mark = "Rear-end")
hotspots_rear_end <- DRHotspots_k_n(X = SampleMarkedPattern, rel_probs = rel_probs_rear_end, 
k = 1, n = 30)
PlotHotspots(X = SampleMarkedPattern, hotspots = hotspots_rear_end)
}
}
