% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleMarkedPattern}
\alias{SampleMarkedPattern}
\title{Marked point pattern on a road network simulating traffic accidents}
\format{An object of class \code{lpp} (inherits from \code{ppx}) of length 3.}
\usage{
SampleMarkedPattern
}
\description{
A simulated \code{lpp} object representing traffic accidents lying on a road structure. The pattern is marked, with the following marks: \code{Collision}, \code{Bicycle}, \code{Car}, \code{Lorry}, \code{Motorcycle}, \code{Public.bus}, \code{Private.bus} and \code{Van}. The mark \code{Collision} is a factor, with the following possible values: \code{Crossing}, \code{Fixed-object}, \code{Rear-end}, \code{Run-off-road}, \code{Run-over}, \code{Side}. The remaining marks are numeric and binary
}
\keyword{datasets}
